refactor: remove snapshot
