+++ src/mask/custom-elements.ts
@@ -11,8 +11,10 @@ import '@/mask/plugin-host/enable.js';
 // custom elements
 import '@/mask/custom-elements/PageInspector.js';
 import '@/mask/custom-elements/CalendarWidget.js';
-import '@/mask/custom-elements/DecryptedPost.js';
-import '@/mask/custom-elements/PostInspector.js';
+import '@/mask/custom-elements/DecryptedPostInspector.js';
+
+// no plugin avaiable for post inspector
+// import '@/mask/custom-elements/PostInspector.js';
 
 import { setPluginDebuggerMessages } from '@/mask/message-host/index.js';
 import { env } from '@/constants/env.js';
