+++ src/mask/custom-elements/DecryptedPostInspector.tsx
