+++ src/mask/plugin-host/register.ts
@@ -1,4 +1,2 @@
-import '@masknet/plugin-debugger/register';
 import '@masknet/plugin-redpacket/register';
 import '@masknet/plugin-calendar/register';
-import '@masknet/plugin-snapshot/register';
